Move overview query state into overview module
