+++ web/features/deployments/detail/overview/README.md
@@ -4,9 +4,8 @@ Overview route for release summary, environment status, and access summaries on 
 
 ## External Modules
 
-| Module                                | Why this module uses it                                                    |
-| ------------------------------------- | -------------------------------------------------------------------------- |
-| `../../route-state`                   | Reads the route app instance identity for overview links.                  |
-| `../../shared/components/empty-state` | Reuses deployment empty/error state presentation for the overview route.   |
-| `../../shared/domain/runtime-status`  | Reuses deployment status rules to compute overview counts.                 |
-| `../state`                            | Reads detail-owned overview query data before composing overview sections. |
+| Module                                | Why this module uses it                                                       |
+| ------------------------------------- | ----------------------------------------------------------------------------- |
+| `../../route-state`                   | Reads the route app instance identity for overview query, links, and actions. |
+| `../../shared/components/empty-state` | Reuses deployment empty/error state presentation for the overview route.      |
+| `../../shared/domain/runtime-status`  | Reuses deployment status rules to compute overview counts.                    |
