+++ web/features/deployments/detail/overview/index.tsx
@@ -6,10 +6,10 @@ import Link from '@/next/link'
 import { deploymentRouteAppInstanceIdAtom } from '../../route-state'
 import { DeploymentStateMessage } from '../../shared/components/empty-state'
 import { hasRuntimeInstanceDeployment } from '../../shared/domain/runtime-status'
-import { deploymentDetailOverviewQueryAtom } from '../state'
 import { AccessStatusSection, AccessStatusSectionSkeleton, ApiTokenSummarySection, ApiTokenSummarySectionSkeleton } from './access-summary/access-status-section'
 import { EnvironmentStrip, EnvironmentStripSkeleton } from './environment-status/environment-strip'
 import { ReleaseHero, ReleaseHeroSkeleton } from './release-summary/release-hero'
+import { deploymentOverviewQueryAtom } from './state'
 
 function OverviewLayout({ children }: { children: React.ReactNode }) {
   return (
@@ -63,7 +63,7 @@ function OverviewLoadingSkeleton() {
 
 export function DeploymentOverview() {
   const { t } = useTranslation('deployments')
-  const overviewQuery = useAtomValue(deploymentDetailOverviewQueryAtom)
+  const overviewQuery = useAtomValue(deploymentOverviewQueryAtom)
   const overview = overviewQuery.data
 
   if (overviewQuery.isLoading)
